Get rdoc rake task working
Require rdoc/task as rake/rdoctask was removed in Rake 10.  Use
hanna itself when for the rdoc, as dogfooding is good.

As rdoc didn't actually generate useful output probably due to
encoding issues, replace non-ascii characters in main file so
that it actually runs.

Add a clean task that cleans the rdoc directory and any generated
.gem files, and change the gitignore file to ignore rdoc instead
of doc.