Use parser gem for syntax highlighting of verbatim sections that use ruby code
The default rdoc parser is very basic and doesn't highlight all
valid ruby code, and highlights code that is not valid ruby.
This uses the parser gem to parse the text.  If the parse is
successful, the text represents valid ruby and should be
highlighted.

If the parser gem is not installed, then no changes are made.