Change syntax highlighting color for comments and strings
Previously, comments were highlighted in a dark red, which was
hard to read against the the dark background.  Lighten it up
so it is easier to read.

Ruby strings were previously not highlighted differently than
the default text, this uses a different shade that is hopefully
still easy to read.

I think this is an improvement, but I'm pretty bad with colors,
so if anyone could come up with a better color scheme (preferrably
with a light background), I'd appreciate it.

Bump version to 0.4.1.