Remove target="docwin" from links, Bump version to 1.0.1
Having target on every link breaks expected behavior when a
specific frame is linked to.  It's not actually needed to set
target="docwin" on links at all, because the three navigation
frames all use target="base" in the base tag, which sets the
default target for all links in that frame.